Initial commit for `The Forge`
This is a project to house all little projects too small for their own
repo. Initial musings are AI-related projects.

- yt-summarizer retrieves all videos in a youtube playlist, downloads
the audio, uses OpenAI Whisper to transcribe them, and then asks ChatGPT
to summarize them
- god is a library for delegating code-writing tasks to chatgpt. You can
either call a function and have the function created dynamically. Or,
you can define a function and have chatGPT write the func given the
signature and docstring.